Floating point handling in UniformSnapshot

+++ dataset/GitHub_Java/dropwizard.metrics/metrics-core/src/main/java/com/codahale/metrics/WeightedSnapshot.java
@@ -99,7 +99,7 @@ public class WeightedSnapshot extends Snapshot {
             return values[values.length - 1];
         }
 
-        return values[posx];
+        return values[(int) posx];
     }
 
     /**
